requirements.txt
torch>=2.0
torchvision>=0.15
numpy
opencv-python
Pillow
albumentations>=1.3
tqdm
scikit-learn
